moved API functionalities into app.ts
